+++ src/main/java/com/github/hcsp/inheritance/Animals.java
@@ -1,0 +1,24 @@
+package com.github.hcsp.inheritance;
+
+public class Animals {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Animals(String name) {
+        this.name = name;
+    }
+
+    public void sayMyName() {
+        System.out.println("我的名字是" + name);
+    }
+
+
+
+}
